feat: comments (#9)
* feat: email

* feat: verification code and register

* feat: add pwd verification code

* feat: move redis operations from email to user

* feat:  forget password

* feat: critical user info

* feat: critical user info rpc update database

* no message

* style: format api comments

* feat: PR fix

* fix: modify variable name & swap the order of sending email and setting cache

* feat: get critical user info

* feat: like

* fix: swap limit and offset parameters position

* feat: comment & delete comment

* feat: comment

* fix: redis pipeline bug

* feat: db query by order

* fix: correct the order of returning inner floor counts slice, let it be same as the root ids' order

* feat: add like count to comment db and cache

* refactor: delete annotation

+++ src/service/like/model/like_model.go
@@ -6,11 +6,11 @@ type Like struct {
 	gorm.Model
 	MessageID int64 `gorm:"index:idx_like_message_uid_type,priority:3,unique"`
 	Uid       int64 `gorm:"index:idx_like_message_uid_type,priority:1,unique"`
-	TypeID    uint8 `gorm:"index:idx_like_message_uid_type,priority:2,unique"`
+	TypeID    int   `gorm:"index:idx_like_message_uid_type,priority:2,unique"`
 }
 
 type LikeCount struct {
-	TypeID    uint8 `gorm:"index:idx_like_message_type,priority:2"`
+	TypeID    int   `gorm:"index:idx_like_message_type,priority:2"`
 	MessageID int64 `gorm:"index:idx_like_message_type,priority:1"`
 	Count     int
 }
